Almost rendering

+++ src/World.h
@@ -5,7 +5,7 @@
 #include "OscController.h"
 
 struct World {
-	ci::ivec2 windowSize;
+	ci::vec2 windowSize;
 	std::shared_ptr<OscController> oscController;
 	float dt;
 	float time;
